# go-service-template/api/product/product.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Product Service",
    "version": "1.0",
    "contact": {
      "name": "Pharmacity Internal Team",
      "url": "https://github.com/Pharmacity-JSC"
    },
    "license": {
      "name": "BSD 3-Clause License",
      "url": "https://github.com/grpc-ecosystem/grpc-gateway/blob/master/LICENSE.txt"
    }
  },
  "tags": [
    {
      "name": "Product"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/boilerplate-svc/api/v1/kafka/check": {
      "get": {
        "operationId": "Product_CheckKafka",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/productCheckKafkaResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "Product"
        ]
      }
    },
    "/boilerplate-svc/api/v1/products": {
      "get": {
        "operationId": "Product_ListProducts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/productListProductsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "description": "@gotags: form:\"page,default=1\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "description": "@gotags: form:\"limit,default=10\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "sort_order",
            "description": "@gotags: form:\"sort_order,default=DESC\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sort_by",
            "description": "@gotags: form:\"sort_by,default=id\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sku",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Product"
        ]
      },
      "post": {
        "operationId": "Product_CreateProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/productCreateProductRequest"
            }
          }
        ],
        "tags": [
          "Product"
        ]
      }
    },
    "/boilerplate-svc/api/v1/products/cache/{sku}": {
      "get": {
        "operationId": "Product_GetProductFromCache",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/productGetProductFromCacheResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sku",
            "description": "@gotags: json:\"-\" path:\"sku\"",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Product"
        ]
      }
    },
    "/boilerplate-svc/api/v1/products/export": {
      "get": {
        "operationId": "Product_ExportProducts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protobufFileInfo"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "export_type",
            "description": "filter request",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Product"
        ]
      }
    },
    "/boilerplate-svc/api/v1/products/{uid}": {
      "get": {
        "operationId": "Product_GetProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/productGetProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uid",
            "description": "@gotags: json:\"-\" path:\"uid\"",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Product"
        ]
      },
      "delete": {
        "operationId": "Product_DeleteProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uid",
            "description": "@gotags: json:\"-\" path:\"uid\"",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Product"
        ]
      },
      "put": {
        "operationId": "Product_UpdateProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uid",
            "description": "@gotags: json:\"-\" path:\"uid\"",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductUpdateProductBody"
            }
          }
        ],
        "tags": [
          "Product"
        ]
      }
    }
  },
  "definitions": {
    "ProductUpdateProductBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "sku": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/templateproductStatus"
        },
        "price": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "productCheckKafkaResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "productCreateProductRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "sku": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/templateproductStatus"
        },
        "price": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "productGetProductFromCacheResponse": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string"
        },
        "data": {
          "type": "string",
          "title": "JSON string of product"
        },
        "from_cache": {
          "type": "boolean"
        }
      }
    },
    "productGetProductResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "uid": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "sku": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/templateproductStatus"
        },
        "price": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "productListProductsResponse": {
      "type": "object",
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/productGetProductResponse"
          }
        },
        "pagination": {
          "$ref": "#/definitions/protobufPagination"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "protobufFileInfo": {
      "type": "object",
      "properties": {
        "file_name": {
          "type": "string"
        },
        "file_data": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "protobufPagination": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "total_page": {
          "type": "integer",
          "format": "int32"
        },
        "total_records": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "templateproductStatus": {
      "type": "string",
      "enum": [
        "STATUS_UNSPECIFIED",
        "STATUS_ACTIVE",
        "STATUS_INACTIVE",
        "STATUS_ARCHIVED"
      ],
      "default": "STATUS_UNSPECIFIED"
    }
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "description": "Bearer Token",
      "name": "Authorization",
      "in": "header"
    }
  }
}