# go-service-template/api/product/product_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v1.0.1
// - protoc             v5.26.1
// source: api/product/product.proto

package productpb

import (
	context "context"
	logx "github.com/zeromicro/go-zero/core/logx"
	rest "github.com/zeromicro/go-zero/rest"
	errors "go-service-template/helper/errors"
	protobuf "go-service-template/helper/pmcpb/protobuf"
	middleware "go-service-template/helper/server/http/middleware"
	request "go-service-template/helper/server/http/request"
	response "go-service-template/helper/server/http/response"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
	http "net/http"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the go-zero package it is being compiled against.
var _ = new(context.Context)
var _ = new(response.Response)
var _ = http.StatusOK
var _ = rest.RestConf{}
var _ = request.MaxFileNumber
var _ = errors.DefaultMsg
var _ = middleware.Name
var _ = logx.ErrLogFileClosed

const OperationProductCheckKafka = "/go.service.template.product.Product/CheckKafka"
const OperationProductCreateProduct = "/go.service.template.product.Product/CreateProduct"
const OperationProductDeleteProduct = "/go.service.template.product.Product/DeleteProduct"
const OperationProductExportProducts = "/go.service.template.product.Product/ExportProducts"
const OperationProductGetProduct = "/go.service.template.product.Product/GetProduct"
const OperationProductGetProductFromCache = "/go.service.template.product.Product/GetProductFromCache"
const OperationProductListProducts = "/go.service.template.product.Product/ListProducts"
const OperationProductUpdateProduct = "/go.service.template.product.Product/UpdateProduct"

type ProductHTTPServer interface {
	CheckKafka(context.Context, *CheckKafkaRequest) (*CheckKafkaResponse, error)
	CreateProduct(context.Context, *CreateProductRequest) (*emptypb.Empty, error)
	DeleteProduct(context.Context, *DeleteProductRequest) (*emptypb.Empty, error)
	ExportProducts(context.Context, *ExportProductsRequest) (*protobuf.FileInfo, error)
	GetProduct(context.Context, *GetProductRequest) (*GetProductResponse, error)
	GetProductFromCache(context.Context, *GetProductFromCacheRequest) (*GetProductFromCacheResponse, error)
	ListProducts(context.Context, *ListProductsRequest) (*ListProductsResponse, error)
	UpdateProduct(context.Context, *UpdateProductRequest) (*emptypb.Empty, error)
}

func RegisterProductHTTPServer(s *rest.Server, srv ProductHTTPServer) []rest.Route {
	var routes = []rest.Route{
		{
			Method:  "GET",
			Path:    "/boilerplate-svc/api/v1/products",
			Handler: _Product_ListProducts0_HTTP_Handler(srv),
		},
		{
			Method:  "GET",
			Path:    "/boilerplate-svc/api/v1/products/:uid",
			Handler: _Product_GetProduct0_HTTP_Handler(srv),
		},
		{
			Method:  "POST",
			Path:    "/boilerplate-svc/api/v1/products",
			Handler: _Product_CreateProduct0_HTTP_Handler(srv),
		},
		{
			Method:  "PUT",
			Path:    "/boilerplate-svc/api/v1/products/:uid",
			Handler: _Product_UpdateProduct0_HTTP_Handler(srv),
		},
		{
			Method:  "DELETE",
			Path:    "/boilerplate-svc/api/v1/products/:uid",
			Handler: _Product_DeleteProduct0_HTTP_Handler(srv),
		},
		{
			Method:  "GET",
			Path:    "/boilerplate-svc/api/v1/products/export",
			Handler: _Product_ExportProducts0_HTTP_Handler(srv),
		},
		{
			Method:  "GET",
			Path:    "/boilerplate-svc/api/v1/products/cache/:sku",
			Handler: _Product_GetProductFromCache0_HTTP_Handler(srv),
		},
		{
			Method:  "GET",
			Path:    "/boilerplate-svc/api/v1/kafka/check",
			Handler: _Product_CheckKafka0_HTTP_Handler(srv),
		},
	}
	return routes
}

func _Product_ListProducts0_HTTP_Handler(srv ProductHTTPServer) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var in = &ListProductsRequest{}
		if binder, ok := interface{}(in).(request.RequestBinder); ok {
			if err := binder.Bind(r); err != nil {
				response.Error(r.Context(), w, errors.BadRequest(err))
				return
			}
		} else if err := request.Parse(r, in); err != nil {
			response.Error(r.Context(), w, errors.BadRequest(err))
			return
		}

		resp, err := srv.ListProducts(r.Context(), in)
		if err != nil {
			response.Error(r.Context(), w, err)
			return
		}

		if err := response.Write(r.Context(), w, r, resp); err != nil {
			logx.WithContext(r.Context()).Error(err)
		}
	}
}

func _Product_GetProduct0_HTTP_Handler(srv ProductHTTPServer) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var in = &GetProductRequest{}
		if binder, ok := interface{}(in).(request.RequestBinder); ok {
			if err := binder.Bind(r); err != nil {
				response.Error(r.Context(), w, errors.BadRequest(err))
				return
			}
		} else if err := request.Parse(r, in); err != nil {
			response.Error(r.Context(), w, errors.BadRequest(err))
			return
		}

		resp, err := srv.GetProduct(r.Context(), in)
		if err != nil {
			response.Error(r.Context(), w, err)
			return
		}

		if err := response.Write(r.Context(), w, r, resp); err != nil {
			logx.WithContext(r.Context()).Error(err)
		}
	}
}

func _Product_CreateProduct0_HTTP_Handler(srv ProductHTTPServer) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var in = &CreateProductRequest{}
		if binder, ok := interface{}(in).(request.RequestBinder); ok {
			if err := binder.Bind(r); err != nil {
				response.Error(r.Context(), w, errors.BadRequest(err))
				return
			}
		} else if err := request.Parse(r, in); err != nil {
			response.Error(r.Context(), w, errors.BadRequest(err))
			return
		}

		resp, err := srv.CreateProduct(r.Context(), in)
		if err != nil {
			response.Error(r.Context(), w, err)
			return
		}

		if err := response.Write(r.Context(), w, r, resp); err != nil {
			logx.WithContext(r.Context()).Error(err)
		}
	}
}

func _Product_UpdateProduct0_HTTP_Handler(srv ProductHTTPServer) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var in = &UpdateProductRequest{}
		if binder, ok := interface{}(in).(request.RequestBinder); ok {
			if err := binder.Bind(r); err != nil {
				response.Error(r.Context(), w, errors.BadRequest(err))
				return
			}
		} else if err := request.Parse(r, in); err != nil {
			response.Error(r.Context(), w, errors.BadRequest(err))
			return
		}

		resp, err := srv.UpdateProduct(r.Context(), in)
		if err != nil {
			response.Error(r.Context(), w, err)
			return
		}

		if err := response.Write(r.Context(), w, r, resp); err != nil {
			logx.WithContext(r.Context()).Error(err)
		}
	}
}

func _Product_DeleteProduct0_HTTP_Handler(srv ProductHTTPServer) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var in = &DeleteProductRequest{}
		if binder, ok := interface{}(in).(request.RequestBinder); ok {
			if err := binder.Bind(r); err != nil {
				response.Error(r.Context(), w, errors.BadRequest(err))
				return
			}
		} else if err := request.Parse(r, in); err != nil {
			response.Error(r.Context(), w, errors.BadRequest(err))
			return
		}

		resp, err := srv.DeleteProduct(r.Context(), in)
		if err != nil {
			response.Error(r.Context(), w, err)
			return
		}

		if err := response.Write(r.Context(), w, r, resp); err != nil {
			logx.WithContext(r.Context()).Error(err)
		}
	}
}

func _Product_ExportProducts0_HTTP_Handler(srv ProductHTTPServer) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var in = &ExportProductsRequest{}
		if binder, ok := interface{}(in).(request.RequestBinder); ok {
			if err := binder.Bind(r); err != nil {
				response.Error(r.Context(), w, errors.BadRequest(err))
				return
			}
		} else if err := request.Parse(r, in); err != nil {
			response.Error(r.Context(), w, errors.BadRequest(err))
			return
		}

		resp, err := srv.ExportProducts(r.Context(), in)
		if err != nil {
			response.Error(r.Context(), w, err)
			return
		}

		if err := response.Write(r.Context(), w, r, resp); err != nil {
			logx.WithContext(r.Context()).Error(err)
		}
	}
}

func _Product_GetProductFromCache0_HTTP_Handler(srv ProductHTTPServer) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var in = &GetProductFromCacheRequest{}
		if binder, ok := interface{}(in).(request.RequestBinder); ok {
			if err := binder.Bind(r); err != nil {
				response.Error(r.Context(), w, errors.BadRequest(err))
				return
			}
		} else if err := request.Parse(r, in); err != nil {
			response.Error(r.Context(), w, errors.BadRequest(err))
			return
		}

		resp, err := srv.GetProductFromCache(r.Context(), in)
		if err != nil {
			response.Error(r.Context(), w, err)
			return
		}

		if err := response.Write(r.Context(), w, r, resp); err != nil {
			logx.WithContext(r.Context()).Error(err)
		}
	}
}

func _Product_CheckKafka0_HTTP_Handler(srv ProductHTTPServer) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var in = &CheckKafkaRequest{}
		if binder, ok := interface{}(in).(request.RequestBinder); ok {
			if err := binder.Bind(r); err != nil {
				response.Error(r.Context(), w, errors.BadRequest(err))
				return
			}
		} else if err := request.Parse(r, in); err != nil {
			response.Error(r.Context(), w, errors.BadRequest(err))
			return
		}

		resp, err := srv.CheckKafka(r.Context(), in)
		if err != nil {
			response.Error(r.Context(), w, err)
			return
		}

		if err := response.Write(r.Context(), w, r, resp); err != nil {
			logx.WithContext(r.Context()).Error(err)
		}
	}
}